refactor toggle hype block size to shared component
