--- packages/lib/modules/chains/hyperevm/index.ts
@@ -1,3 +1,0 @@
-export * from './useIsHyperEvm'
-export * from './useIsUsingBigBlocks'
-export * from './useSetUsingBigBlocks'
