--- packages/lib/modules/chains/hyperevm/useIsHyperEvm.ts
@@ -1,8 +1,0 @@
-import { ChainId } from '@balancer/sdk'
-import { usePublicClient } from 'wagmi'
-
-export const useIsHyperEvm = () => {
-  const publicClient = usePublicClient()
-
-  return publicClient?.chain.id === ChainId.HYPER_EVM
-}
