+++ packages/lib/modules/pool/actions/create/modal/PoolCreationModal.tsx
@@ -3,7 +3,6 @@ import { Modal, ModalBody, ModalCloseButton, ModalContent, ModalProps } from '@c
 import { RefObject, useRef } from 'react'
 import { TransactionModalHeader } from '@repo/lib/shared/components/modals/TransactionModalHeader'
 import { SuccessOverlay } from '@repo/lib/shared/components/modals/SuccessOverlay'
-
 import { useBreakpoints } from '@repo/lib/shared/hooks/useBreakpoints'
 import { VStack, Button, HStack, Text } from '@chakra-ui/react'
 import { getPoolPath } from '@repo/lib/modules/pool/pool.utils'
@@ -15,21 +14,13 @@ import { ActionModalFooter } from '@repo/lib/shared/components/modals/ActionModa
 import { Address } from 'viem'
 import { useIsPoolInitialized } from '@repo/lib/modules/pool/queries/useIsPoolInitialized'
 import { getChainId, getChainName } from '@repo/lib/config/app.config'
-import { BalAlert } from '@repo/lib/shared/components/alerts/BalAlert'
 import { useShouldBatchTransactions } from '@repo/lib/modules/web3/safe.hooks'
-import {
-  useIsUsingBigBlocks,
-  useSetUsingBigBlocks,
-  useIsHyperEvm,
-} from '@repo/lib/modules/chains/hyperevm'
-import { LabelWithIcon } from '@repo/lib/shared/components/btns/button-group/LabelWithIcon'
-import { useUserAccount } from '@repo/lib/modules/web3/UserAccountProvider'
-import { ExternalLink } from 'react-feather'
-import { BalAlertLink } from '@repo/lib/shared/components/alerts/BalAlertLink'
 import { usePoolCreation } from './PoolCreationProvider'
 import { usePoolCreationForm } from '../PoolCreationFormProvider'
 import { PoolType } from '@balancer/sdk'
 import { PoolSummary } from './PoolSummary'
+import { ToggleHyperBlockSize } from './ToggleHyperBlockSize'
+import { useHyperEvm } from '@repo/lib/modules/chains/hyperevm/useHyperEvm'
 
 type Props = {
   isOpen: boolean
@@ -46,10 +37,8 @@ export function PoolCreationModal({
 }: Props & Omit<ModalProps, 'children'>) {
   const initialFocusRef = useRef(null)
   const { isDesktop } = useBreakpoints()
-  const { isConnected } = useUserAccount()
   const { transactionSteps, initPoolTxHash, urlTxHash } = usePoolCreation()
   const { network, poolType, resetPoolCreationForm } = usePoolCreationForm()
-  const { isUsingBigBlocks } = useIsUsingBigBlocks()
 
   const [poolAddress] = useLocalStorage<Address | undefined>(
     LS_KEYS.PoolCreation.Address,
@@ -84,15 +73,13 @@ export function PoolCreationModal({
 
   const isSuccess = !!isPoolInitialized
 
-  const isHyperEvm = useIsHyperEvm()
-  const shouldUseBigBlocks =
-    isConnected && isHyperEvm && !isUsingBigBlocks && transactionSteps.currentStepIndex === 0 // big blocks only necessary to deploy pool contract
-  const shouldUseSmallBlocks =
-    isConnected && isHyperEvm && !!isUsingBigBlocks && transactionSteps.currentStepIndex > 0 // small blocks faster for non contract deployment txs
-  const shouldToggleBlockSize = shouldUseBigBlocks || shouldUseSmallBlocks
-
-  const { setUsingBigBlocks, isSetUsingBigBlocksPending, setUsingBigBlocksError } =
-    useSetUsingBigBlocks()
+  const {
+    shouldUseBigBlocks,
+    shouldToggleBlockSize,
+    setUsingBigBlocks,
+    isSetUsingBigBlocksPending,
+    setUsingBigBlocksError,
+  } = useHyperEvm({ isContractDeploymentStep: transactionSteps.currentStepIndex === 0 })
 
   return (
     <Modal
@@ -163,43 +150,12 @@ export function PoolCreationModal({
         </ModalBody>
 
         {shouldToggleBlockSize ? (
-          <VStack marginTop="2" paddingLeft="6" paddingRight="6" spacing="3" width="full">
-            <Button
-              isDisabled={isSetUsingBigBlocksPending}
-              isLoading={isSetUsingBigBlocksPending}
-              onClick={() => setUsingBigBlocks(shouldUseBigBlocks)}
-              size="lg"
-              variant="primary"
-              w="full"
-            >
-              <LabelWithIcon icon="sign">
-                Switch to {shouldUseBigBlocks ? 'big' : 'small'} blocks
-              </LabelWithIcon>
-            </Button>
-            {setUsingBigBlocksError ? (
-              <BalAlert content={setUsingBigBlocksError.message} status="error" title="Error:" />
-            ) : (
-              <BalAlert
-                content={
-                  <Text color="font.primaryGradient">
-                    {shouldUseBigBlocks
-                      ? `The HyperEVM network requires your account switch to using big blocks to deploy contracts.`
-                      : `Your HyperEvm account is using big blocks. Switch to small blocks for faster transaction speeds.`}{' '}
-                    <BalAlertLink
-                      alignItems="center"
-                      display="inline-flex"
-                      href="https://hyperliquid.gitbook.io/hyperliquid-docs/for-developers/hyperevm/dual-block-architecture"
-                      isExternal
-                    >
-                      <span>See more information</span>
-                      <ExternalLink size={16} style={{ marginLeft: 2, verticalAlign: 'middle' }} />
-                    </BalAlertLink>
-                  </Text>
-                }
-                status="info"
-              />
-            )}
-          </VStack>
+          <ToggleHyperBlockSize
+            isSetUsingBigBlocksPending={isSetUsingBigBlocksPending}
+            setUsingBigBlocks={setUsingBigBlocks}
+            setUsingBigBlocksError={setUsingBigBlocksError}
+            shouldUseBigBlocks={shouldUseBigBlocks}
+          />
         ) : (
           <ActionModalFooter
             currentStep={transactionSteps.currentStep}
